Added gec manager
