+++ app/grammar_checker.py
@@ -1,5 +1,6 @@
 import language_tool_python as lt
-from happytransformer import HappyTextToText, TTSettings
+from app.gec.gector_large import predict as gl
+import gec.language_tool.predict as lt
 
 ###
 # Errors detected by the grammar checker:
@@ -14,104 +15,23 @@ from happytransformer import HappyTextToText, TTSettings
 ###
 
 class GrammarChecker:
-    # Can run in public API mode or local mode (.jar server)
+    # Language Tool can run in public API mode or local mode (.jar server)
+    _gec_models={
+        "GECTOR_2024",
+        "LT_API",
+        "LT_SERVER"
+    }
 
-    def __init__(self, language='en-US', public_api=False, filter_errors=True):
-        if public_api:
+    def __init__(self, gec_model="GECTOR_2024"):
+        if gec_model=="LT_API":
             print("Using language tool public API.")
-            self.__tool = lt.LanguageToolPublicAPI(language)
-        else:
-            print("Using language tool local.")
-            self.__tool = lt.LanguageTool(language, config={
-                'cacheSize': 1024, # Improves performance
-                'pipelineCaching': True,
-            })
+            self.__tool = lt.LT_Checker(gec_model=gec_model)
+        elif gec_model=="LT_SERVER":
+            self.__tool = lt.LT_Checker(gec_model=gec_model)
+        elif gec_model=="GECTOR_2024":
+            print("Using GECToR-2024")
+            self.__tool = gl
 
-        self.__categories = [
-            'COLLOCATIONS',
-            'CONFUSED_WORDS',
-            'GRAMMAR',
-            'MISC',  # Partially included
-            'REDUNDANCY',
-            'REPETITIONS_STYLE'
-        ]
-        self.__avoid_rule_ids = [
-            'EN_COMPOUNDS',
-            'EN_CONSISTENT_APOS',
-            'EN_SIMPLE_REPLACE',
-            'REP_PASSIVE_VOICE',
-            'REP_THANK_YOU_FOR'
-        ]
-        self.__filter_errors = filter_errors
-        self.__t5_model = HappyTextToText("T5", "vennify/t5-base-grammar-correction")
+    def correct_sentences(self, sentences):
+        self.__tool.correct_sentences(sentences)
 
-
-    # Given a text (string)
-    # Return a list of matches (dicts)
-    def check(self, text: str):
-        matches = self.__tool.check(text)
-        matches_dicts = [{
-            'ruleId': match.ruleId,
-            'message': match.message,
-            'replacements': match.replacements,
-            'offsetInContext': match.offsetInContext,
-            'context': match.context,
-            'offset': match.offset,
-            'errorLength': match.errorLength,
-            'category': match.category,
-            'ruleIssueType': match.ruleIssueType,
-            'sentence': match.sentence
-        } for match in matches]
-
-        if self.__filter_errors:
-            matches_dicts = self.filter_errors(matches_dicts)
-        
-        return matches_dicts
-
-
-    # Given a text (string)
-    # Return a corrected text (string)
-    def correct(self, text: str):
-        matches = self.__tool.check(text)
-        return lt.utils.correct(text, matches)
-
-
-    # Given a list of sentences (strings)
-    # Return a list of matches for all sentences (dicts)
-    def check_sentences(self, sentences: list):
-        all_matches = []
-        for sentence in sentences:
-            matches = self.check(sentence)
-            if matches:
-                all_matches.extend(matches)
-
-        sorted_matches = sorted(all_matches, key=lambda x: x['ruleId'])
-
-        return sorted_matches
-
-
-    # Given a list of sentences (strings)
-    # Return a list of corrected sentences (strings)
-    def correct_sentences(self, sentences: list):
-        corrected_sentences = []
-        for sentence in sentences:
-            corrected = self.correct(sentence)
-            corrected_sentences.extend(corrected)
-        return corrected_sentences
-    
-
-    # Given a list of matches (dicts)
-    # Return a list of matches with the category and ruleId filtered
-    def filter_errors(self, matches: list):
-        filtered_matches = [match for match in matches if match['category'] in self.__categories and match['ruleId'] not in self.__avoid_rule_ids]
-        return filtered_matches
-
-    def t5_check_sentence(self, sentence: str):
-        args = TTSettings(num_beams=5, min_length=1)
-        result = self.__t5_model.generate_text(f"grammar: {sentence}", args=args)
-        return result.text
-    
-
-    # Close the language tool (for background .jar server)
-    def close(self):
-        self.__tool.close()
